+++ compiler3.cpp
@@ -8,7 +8,7 @@ Token look_ahead(std::vector<Token> &tokens, int &amount);
 Token look_behind(std::vector<Token> &tokens, int &amount);
 
 void syntaxanalysis(std::vector<Token> &tokens) {
-    for(uint16_t cur_ind = 0; cur_ind < tokens.size(); cur_ind++){
+    for(size_t cur_ind = 0; cur_ind < tokens.size(); cur_ind++){
         std::cout << "Type: '" << tokens[cur_ind].type << "' Subtype: '" << tokens[cur_ind].subtype << "' Value: '" << tokens[cur_ind].value << "'" << std::endl;
         switch (tokens[cur_ind].type){
         case Tokens::KEYWORD:
@@ -34,10 +34,10 @@ void syntaxanalysis(std::vector<Token> &tokens) {
     }
 }
 
-Token look_ahead(std::vector<Token> &tokens, int &amount){
+Token look_ahead(std::vector<Token> &tokens, size_t &amount){
     return tokens[amount + 1];
 }
 
-Token look_behind(std::vector<Token> &tokens, int &amount){
+Token look_behind(std::vector<Token> &tokens, size_t &amount){
     return tokens[amount - 1];
 }
